+++ src/Controller/Admin/DocumentTypeCrudController.php
@@ -28,6 +28,7 @@ class DocumentTypeCrudController extends AbstractCrudController
     public function configureFields(string $pageName): iterable
     {
         yield TextField::new('name', 'Name')
-            ->setRequired(true);
+            ->setRequired(true)
+        ;
     }
 }