+++ src/Repository/CaseEntityRepository.php
@@ -30,7 +30,8 @@ class CaseEntityRepository extends ServiceEntityRepository
             ->orderBy('c.createdAt', 'DESC')
             ->setMaxResults(1)
             ->getQuery()
-            ->getOneOrNullResult();
+            ->getOneOrNullResult()
+        ;
     }
 
     public function findReadyCasesWithoutActiveAgendaByBoard(Board $board)