+++ tests/Service/CaseManagerTest.php
@@ -6,11 +6,13 @@ use App\Entity\CaseEntity;
 use App\Entity\Municipality;
 use App\Repository\CaseEntityRepository;
 use App\Service\CaseManager;
+use App\Service\WorkflowService;
 use PHPUnit\Framework\TestCase;
 
 class CaseManagerTest extends TestCase
 {
     private $mockCaseRepository;
+    private $mockWorkflowService;
     private $caseManager;
 
     protected function setUp(): void
@@ -18,8 +20,12 @@ class CaseManagerTest extends TestCase
         parent::setUp();
 
         $this->mockCaseRepository = $this->createMock(CaseEntityRepository::class);
+        $this->mockWorkflowService = $this->createMock(WorkflowService::class);
 
-        $this->caseManager = new CaseManager($this->mockCaseRepository);
+        $this->caseManager = new CaseManager(
+            $this->mockCaseRepository,
+            $this->mockWorkflowService
+        );
     }
 
     public function testGenerateCaseNumberFirstCase()
@@ -30,7 +36,8 @@ class CaseManagerTest extends TestCase
             ->expects($this->once())
             ->method('findLatestCaseByMunicipality')
             ->with($mockMunicipality)
-            ->willReturn(null);
+            ->willReturn(null)
+        ;
 
         $date = new \DateTime();
         $year = $date->format('Y');
@@ -52,14 +59,16 @@ class CaseManagerTest extends TestCase
             ->expects($this->once())
             ->method('findLatestCaseByMunicipality')
             ->with($mockMunicipality)
-            ->willReturn($mockCase);
+            ->willReturn($mockCase)
+        ;
 
         $mockDate = $this->createMock(\DateTime::class);
 
         $mockCase
             ->expects($this->once())
             ->method('getCreatedAt')
-            ->willReturn($mockDate);
+            ->willReturn($mockDate)
+        ;
 
         $currentDate = new \DateTime();
         $currentYear = $currentDate->format('Y');
@@ -72,7 +81,8 @@ class CaseManagerTest extends TestCase
             ->expects($this->once())
             ->method('format')
             ->with('Y')
-            ->willReturn($previousYear);
+            ->willReturn($previousYear)
+        ;
 
         $actual = $this->caseManager->generateCaseNumber($mockMunicipality);